fix (maybe?) the summary/lobby race condition
